requirements.txt
Django==1.6.1
South==0.8.4
django-crispy-forms==1.4.0
wsgiref==0.1.2
